move extra field parsers to the correct file
